Modify canonical utterance class
