--- plot/generate_utterance.java
@@ -1,28 +1,0 @@
-import java.util.*
-
-public class CanonicalUtterance
-{
-	String path;
-	String value;
-	public CanonicalUtterance() { }
-	public CanonicalUtterance(String path, String value) {
-		this.path = path;
-		this.value = value;
-	}
-
-
-	public String simple_canonical_utterance()
-	{
-
-	}
-
-
-	public String complex_canonical_utterance()
-	{
-
-	}
-
-}
-
-
-
